Start of exercise 8
I have the extent I want, but I can't seem to get this working.  I have the geojson code in a seperate directory, but not all of it will validate, even though I took it DIRECTLY from geojson.io  I have a line, a polygon, and points representing the great lakes, but the code is broken.  I will try to fix another day.  I am out of time.