Update assignment 2
I added a popup function to the choropleth map.  This addition has caused my layer control and legend to disappear.  I added some code for a heatmap.  I think I need to point the heat map code to a https instead of a direct file.  I'll try that after pushing this point layer geojson of parks.